refactor: Reorganize component directory structure
- Move components into feature-based folders (features/, layout/, ui/, icons/)
- Migrate helpers to lib directory
- Move styles and type definitions to root level
- Update all import paths to match new structure
- Remove unused legacy components
- Update ESLint configuration for new structure
