+++ app/layout.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from "next"
-import { Noto_Sans_JP } from "next/font/google"
+import { Noto_Sans_JP, Permanent_Marker } from "next/font/google"
 import StyledComponentsRegistry from "@/lib/registry"
 import Providers from "@/components/Providers"
 import "modern-normalize/modern-normalize.css"
@@ -8,9 +8,16 @@ import "font-awesome/css/font-awesome.css"
 
 const notoSansJP = Noto_Sans_JP({
   subsets: ["latin"],
-  weight: ["400", "700"],
+  weight: ["400", "900"],
   display: "swap",
   variable: "--font-noto-sans-jp",
+})
+
+const permanentMarker = Permanent_Marker({
+  subsets: ["latin"],
+  weight: "400",
+  display: "swap",
+  variable: "--font-permanent-marker",
 })
 
 export const metadata: Metadata = {
@@ -25,7 +32,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="ja">
-      <body className={notoSansJP.className}>
+      <body className={`${notoSansJP.className} ${permanentMarker.variable}`}>
         <StyledComponentsRegistry>
           <Providers>{children}</Providers>
         </StyledComponentsRegistry>
