+++ components/icons/icon.tsx
