+++ package.json
@@ -32,15 +32,16 @@
     "@types/react": "^18.3.11",
     "@types/react-helmet": "6.1.11",
     "@types/styled-components": "5.1.34",
-    "@typescript-eslint/eslint-plugin": "^6.0.0",
-    "@typescript-eslint/parser": "^6.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.44.1",
+    "@typescript-eslint/parser": "^8.44.1",
     "babel-plugin-styled-components": "2.1.4",
-    "eslint": "7.32.0",
+    "eslint": "8.57.1",
+    "eslint-config-next": "^15.5.3",
     "eslint-config-prettier": "8.10.0",
     "eslint-plugin-import": "2.31.0",
     "eslint-plugin-prettier": "3.4.1",
     "eslint-plugin-react": "7.37.4",
-    "eslint-plugin-react-hooks": "4.6.2",
+    "eslint-plugin-react-hooks": "5.2.0",
     "prettier": "2.8.8",
     "styled-components": "6.1.16",
     "textlint": "11.9.1",
@@ -66,7 +67,7 @@
     "format": "prettier --write \"**/*.+(ts|tsx|json|md)\" --ignore-path .gitignore",
     "lint:text": "textlint content/**/index.md",
     "lint:textfix": "textlint --fix content/**/index.md",
-    "lint": "eslint --fix --ext .ts,.tsx --ignore-path .gitignore src",
+    "lint": "eslint --fix --ext .ts,.tsx --ignore-path .gitignore app components lib styles types",
     "test": "tsc -p ./tsconfig.json"
   },
   "packageManager": "pnpm@9.12.3"
