--- src/components/footer.tsx
@@ -1,32 +1,0 @@
-import React from "react"
-import Link from "next/link"
-import styled from "styled-components"
-
-import Hr from "./hr"
-import Container from "./container"
-
-interface Props {
-  author: string
-}
-
-const Footer: React.FC<Props> = ({ author }: Props) => {
-  return (
-    <Container>
-      <StyledFooter>
-        <Hr />
-        <p>コーラとバグが好き</p>
-        <Link href="/profile/">
-          <p>
-            <strong>{author}</strong> on Profile
-          </p>
-        </Link>
-      </StyledFooter>
-    </Container>
-  )
-}
-
-export default Footer
-
-const StyledFooter = styled.div`
-  padding: 2rem;
-`
