--- src/components/lead.tsx
@@ -1,10 +1,0 @@
-import styled from "styled-components"
-
-const Lead = styled.p`
-  margin: 2rem auto;
-  padding: 0rem;
-  line-height: 1.8rem;
-  text-align: left;
-`
-
-export default Lead
